Don't skip configured matchers for exact file names
