+++ integration-tests/__tests__/cli-handles-exact-filenames.test.js
@@ -1,0 +1,72 @@
+/**
+ * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow
+ */
+
+'use strict';
+
+const path = require('path');
+const SkipOnWindows = require('../../scripts/SkipOnWindows');
+const {extractSummary, cleanup, writeFiles} = require('../Utils');
+const runJest = require('../runJest');
+
+const DIR = path.resolve(__dirname, '../cli_accepts_exact_filenames');
+
+SkipOnWindows.suite();
+
+beforeEach(() => cleanup(DIR));
+afterAll(() => cleanup(DIR));
+
+test('CLI accepts exact file names if matchers matched', () => {
+  writeFiles(DIR, {
+    'bar.spec.js': `
+      test('bar', () => {});
+    `,
+    'foo/bar.spec.js': `
+      test('foo', () => {});
+    `,
+    'package.json': JSON.stringify({jest: {testEnvironment: 'node'}}),
+  });
+
+  const result = runJest(DIR, [
+    '-i',
+    '--forceExit',
+    './bar.spec.js',
+    './foo/bar.spec.js',
+  ]);
+
+  expect(result.status).toBe(0);
+
+  const {rest, summary} = extractSummary(result.stderr);
+
+  expect(rest).toMatchSnapshot();
+  expect(summary).toMatchSnapshot();
+  expect(result.stdout).toMatchSnapshot();
+});
+
+test('CLI skips exact file names if no matchers matched', () => {
+  writeFiles(DIR, {
+    'bar.js': `
+      test('bar', () => {);
+    `,
+    'foo/bar.js': `
+      test('foo', () => {);
+    `,
+    'package.json': JSON.stringify({jest: {testEnvironment: 'node'}}),
+  });
+
+  const result = runJest(DIR, [
+    '-i',
+    '--forceExit',
+    './bar.js',
+    './foo/bar.js',
+  ]);
+
+  expect(result.status).toBe(1);
+  expect(result.stdout).toMatch(/No tests found([\S\s]*)3 files checked./);
+  expect(result.stderr).toEqual('');
+});
